Add files via upload
I did assignment 1 and this were the task I needed to complete, and the code for the task is in a zip file 
1) Benford's Law 
Given a collection of naturally occurring numbers - for example, the
length of rivers around the world, enrollment at UAA over the years,
the number of votes for a candidate by precinct – we might expect
these numbers to have nothing in common.
But what if you looked at the distribution of the leading digit of these
numbers? In other words, count the number of times 1 is the leading
digit, 2 is the leading digit, etc. A natural assumption is that all of these
digits are equally likely so there should be about a 11% chance to see
any one value from 1-9 as the leading digit. What we find instead is
quite remarkable. 1 is the leading digit about 30% of the time, and the
probability drops to around 5% for the digit 9. This distribution holds
for all of these examples despite the disparate sources and is called
Benford's Law. One application of this phenomenon is verification of
voting records - if the number of votes by precinct doesn't match the
expected distribution then that raises questions about vote tampering.

2) Chuck-a-luck - 
Chuck-a-luck is a simple dice game for gamblers. You can play it in
some Vegas casinos. To play, the gambler first places a wager, W. The
gambler then picks any number from 1-6. Let n represent the
gambler's number. Next, the gamber rolls three six-sided dice. If none
of the rolls equals n then the gambler loses W. If one of the rolls equals
n then the gambler wins wager W back (even money, 1-1). If two of the
rolls equals n then the gambler wins twice the wager W (2-1). If all
three of the rolls equals n then the gamber wins three times W (3-1).
Write a program that implements a C++ version of Chuck-a-luck. It
should output appropriate messages so the player knows what has
happened. The player should be prompted to input a wager, select a
number, and to initiate rolling the dice. The gambler should start with
$100. The gambler should be allowed to keep playing until he or she
runs out of money (don't allow a wager that is more than the amount
of money owned) or the gambler decides to quit playing. For the
selected number that should be from 1-6, your program should only
allow the gambler to choose a valid number. For example, if 7 or 0 is
entered then the gambler should be prompted to enter the number
again.